fix(toast): fix Vite Fast Refresh warning in ToastContext
Split context object into toastContextDef.ts (non-component exports)
so ToastContext.tsx exports only the ToastProvider component.
Inlines useToasts state directly into ToastProvider, removing the
redundant useToasts.ts file.
