+++ frontend/src/hooks/useToast.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
-import { ToastContext } from '../context/ToastContext';
-import type { ToastContextValue } from '../context/ToastContext';
+import { ToastContext } from '../context/toastContextDef';
+import type { ToastContextValue } from '../context/toastContextDef';
 
 export function useToast(): ToastContextValue {
   const context = useContext(ToastContext);
